+++ capytaine/green_functions/libDelhommeau/src/constants.f90
@@ -16,6 +16,21 @@ MODULE CONSTANTS
   REAL(KIND=PRE), PARAMETER :: LOG_2 = LOG(REAL(2d0, kind=pre))
   COMPLEX(KIND=PRE), PARAMETER :: II = (0, 1)         ! Imaginary unit
 
+  !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+  ! Parameters for different variants of the Green function.
+  ! The actual number should be irrelevant as long as they are different.
+
+  ! Values for "tabulation_grid_shape"
+  integer, parameter :: LEGACY_GRID = 0  ! Nemoh 2
+  integer, parameter :: SCALED_NEMOH3_GRID = 1
+
+  ! Values for "gf_singularities"
+  integer, parameter :: HIGH_FREQ = 0  ! legacy from Nemoh
+  integer, parameter :: LOW_FREQ = 1  ! aka XieDelhommeau
+
+  !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
 contains
 
   PURE LOGICAL FUNCTION is_infinity(x)