+++ pytest/test_bem_green_functions.py
@@ -71,12 +71,12 @@ def test_symmetry_of_the_green_function_infinite_depth(gf):
     k = 1.0
     xi = np.array([0.0, 0.0, -1.0])
     xj = np.array([1.0, 1.0, -2.0])
-    g1, dg1_sym, dg1_antisym = gf.fortran_core.green_wave.wave_part_infinite_depth(
+    g1, dg1 = gf.fortran_core.green_wave.wave_part_infinite_depth(
         xi, xj, k,
         gf.tabulation_grid_shape_index, gf.tabulated_r_range, gf.tabulated_z_range,
         gf.tabulated_integrals, gf.gf_singularities_index
     )
-    g2, dg2_sym, dg2_antisym = gf.fortran_core.green_wave.wave_part_infinite_depth(
+    g2, dg2 = gf.fortran_core.green_wave.wave_part_infinite_depth(
         xj,
         xi,
         k,
@@ -84,8 +84,8 @@ def test_symmetry_of_the_green_function_infinite_depth(gf):
         gf.tabulated_integrals, gf.gf_singularities_index
     )
     assert g1 == approx(g2)
-    assert dg1_sym == approx(dg2_sym)
-    assert dg1_antisym == approx(-dg2_antisym)
+    assert dg1[0:2] == approx(-dg2[0:2])
+    assert dg1[2] == approx(dg2[2])
 
 
 @pytest.mark.parametrize("gf", gfs)
@@ -97,13 +97,13 @@ def test_symmetry_of_the_green_function_finite_depth_no_prony(gf):
     g1, dg1_sym, dg1_antisym = gf.fortran_core.green_wave.wave_part_finite_depth(
         xi, xj, k, depth,
         gf.tabulation_grid_shape_index, gf.tabulated_r_range, gf.tabulated_z_range,
-        gf.tabulated_integrals, gf.gf_singularities_index,
+        gf.tabulated_integrals,
         np.zeros(1), np.zeros(1), 1
     )
     g2, dg2_sym, dg2_antisym = gf.fortran_core.green_wave.wave_part_finite_depth(
         xj, xi, k, depth,
         gf.tabulation_grid_shape_index, gf.tabulated_r_range, gf.tabulated_z_range,
-        gf.tabulated_integrals, gf.gf_singularities_index,
+        gf.tabulated_integrals,
         np.zeros(1), np.zeros(1), 1
     )
     assert g1 == approx(g2)
@@ -121,13 +121,13 @@ def test_symmetry_of_the_green_function_finite_depth(gf):
     g1, dg1_sym, dg1_antisym = gf.fortran_core.green_wave.wave_part_finite_depth(
         xi, xj, k, depth,
         gf.tabulation_grid_shape_index, gf.tabulated_r_range, gf.tabulated_z_range,
-        gf.tabulated_integrals, gf.gf_singularities_index,
+        gf.tabulated_integrals,
         ambda, a, 31
     )
     g2, dg2_sym, dg2_antisym = gf.fortran_core.green_wave.wave_part_finite_depth(
         xj, xi, k, depth,
         gf.tabulation_grid_shape_index, gf.tabulated_r_range, gf.tabulated_z_range,
-        gf.tabulated_integrals, gf.gf_singularities_index,
+        gf.tabulated_integrals,
         ambda, a, 31
     )
     assert g1 == approx(g2)